Move update tests to `CommandBuilder`
